+++ drivers/adc/Kconfig.xec
@@ -1,7 +1,7 @@
 # Microchip XEC ADC configuration
 
 # Copyright (c) 2019 Intel Corporation
-# Copyright (c) 2021 Microchip Technology Inc.
+# Copyright (c) 2023 Microchip Technology Inc.
 # SPDX-License-Identifier: Apache-2.0
 
 config ADC_XEC
@@ -10,10 +10,3 @@ config ADC_XEC
 	depends on DT_HAS_MICROCHIP_XEC_ADC_ENABLED
 	help
 	  Enable ADC driver for Microchip XEC MCU series.
-
-config ADC_XEC_V2
-	bool "Microchip XEC series ADC V2 driver"
-	default y
-	depends on DT_HAS_MICROCHIP_XEC_ADC_V2_ENABLED
-	help
-	  Enable ADC driver for Microchip XEC MEC172x MCU series.
