+++ dts/arm/microchip/mec172xnsz.dtsi
@@ -688,7 +688,7 @@
 			status = "disabled";
 		};
 		adc0: adc@40007c00 {
-			compatible = "microchip,xec-adc-v2";
+			compatible = "microchip,xec-adc";
 			reg = <0x40007c00 0x90>;
 			interrupts = <78 0>, <79 0>;
 			girqs = <17 8>, <17 9>;
