+++ dts/bindings/adc/microchip,xec-adc.yaml
@@ -1,4 +1,5 @@
 # Copyright (c) 2019, Intel Corporation
+# Copyright (c) 2021, Microchip Technology Inc.
 # SPDX-License-Identifier: Apache-2.0
 
 description: Microchip XEC ADC
@@ -16,6 +17,16 @@ properties:
 
   "#io-channel-cells":
     const: 1
+
+  girqs:
+    type: array
+    required: true
+    description: Array of pairs of GIRQ number and bit position
+
+  pcrs:
+    type: array
+    required: true
+    description: ADC PCR register index and bit position
 
   clktime:
     type: int
